#282 created timebuilder.py to seperate general time from creg time
