+++ src/creg/test/test_cregtime_ideaunits_status.py
@@ -1,32 +1,17 @@
 from src.bud.idea import ideaunit_shop
 from src.bud.bud import budunit_shop
 from src.creg.creg import (
-    get_year_road,
     get_time_min_from_dt,
     add_time_creg_ideaunit,
-    time_str,  # "time"
-    creg_week_str,  # "week"
+    creg_week_str,
     get_cregtime_text,
-    c400_leap_str,
-    c400_clean_str,
-    c100_str,
-    yr4_leap_str,
-    yr4_clean_str,
-    get_sun,  # "Sunday"
-    get_mon,  # "Monday"
-    get_tue,  # "Tuesday"
-    get_wed,  # "Wednesday"
-    get_thu,  # "Thursday"
-    get_fri,  # "Friday"
-    get_sat,  # "Saturday"
-    creg_week_str,  # "week"
-    weeks_str,  # f"{get_week()}s"
-    day_str,  # "day"
-    days_str,  # f"{get_day()}s"
-    year_str,
-    years_str,
+    get_wed,
+    get_thu,
+    creg_week_str,
+    weeks_str,
     creg_weekday_ideaunits,
 )
+from src.creg.timebuilder import time_str, day_str, days_str, get_year_road
 from datetime import datetime
